LPS-58897 Move DummyPortalCache to ToolDependencies, prepare to remove package"com.liferay.portal.cache"
