+++ portal-impl/src/com/liferay/portal/tools/ToolDependencies.java
@@ -14,11 +14,12 @@
 
 package com.liferay.portal.tools;
 
-import com.liferay.portal.cache.DummyPortalCache;
 import com.liferay.portal.cache.key.SimpleCacheKeyGenerator;
 import com.liferay.portal.json.JSONFactoryImpl;
 import com.liferay.portal.kernel.cache.MultiVMPool;
 import com.liferay.portal.kernel.cache.PortalCache;
+import com.liferay.portal.kernel.cache.PortalCacheListener;
+import com.liferay.portal.kernel.cache.PortalCacheListenerScope;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 import com.liferay.portal.kernel.cache.SingleVMPool;
 import com.liferay.portal.kernel.cache.key.CacheKeyGeneratorUtil;
@@ -59,6 +60,8 @@ import com.liferay.registry.RegistryUtil;
 
 import java.io.Serializable;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
@@ -186,6 +189,82 @@ public class ToolDependencies {
 		resourceActionsImpl.afterPropertiesSet();
 
 		resourceActionsUtil.setResourceActions(resourceActionsImpl);
+	}
+
+	private static class DummyPortalCache<K extends Serializable, V>
+		implements PortalCache<K, V> {
+
+		public DummyPortalCache(String portalCacheName) {
+			_portalCacheName = portalCacheName;
+		}
+
+		@Override
+		public V get(K key) {
+			return null;
+		}
+
+		@Override
+		public List<K> getKeys() {
+			return Collections.emptyList();
+		}
+
+		/**
+		 * @deprecated As of 7.0.0, replaced by {@link #getPortalCacheName()}
+		 */
+		@Deprecated
+		@Override
+		public String getName() {
+			return getPortalCacheName();
+		}
+
+		@Override
+		public PortalCacheManager<K, V> getPortalCacheManager() {
+			return null;
+		}
+
+		@Override
+		public String getPortalCacheName() {
+			return _portalCacheName;
+		}
+
+		@Override
+		public void put(K key, V value) {
+		}
+
+		@Override
+		public void put(K key, V value, int timeToLive) {
+		}
+
+		@Override
+		public void registerPortalCacheListener(
+			PortalCacheListener<K, V> portalCacheListener) {
+		}
+
+		@Override
+		public void registerPortalCacheListener(
+			PortalCacheListener<K, V> portalCacheListener,
+			PortalCacheListenerScope portalCacheListenerScope) {
+		}
+
+		@Override
+		public void remove(K key) {
+		}
+
+		@Override
+		public void removeAll() {
+		}
+
+		@Override
+		public void unregisterPortalCacheListener(
+			PortalCacheListener<K, V> portalCacheListener) {
+		}
+
+		@Override
+		public void unregisterPortalCacheListeners() {
+		}
+
+		private final String _portalCacheName;
+
 	}
 
 	private static class TestMultiVMPool implements MultiVMPool {
